rename git to versioning
